Add generator integration doc
Also restructured the top-level README:
1. Removed the pre-release instructions section.
2. Combined import, API, and generator into a single "Getting started with Needle" section.

Moved the section that describes how to invoke generated code into the Generator document, since it deals with generated code, not `NeedleFoundation` API.

Added a section in API document to describe how to bootstrap the Root scope. This section is yet to be filled out.