Kill SourceKitService process before retry parsing (#193)
When parsing times out, it is likely due to SourceKitService hanging. Simply retry again does not seem to resolve the issue. Instead this change force kills the SourceKit service process and re-initializes it. Since SourceKittenFramework does not expose SourceKit service initialization publicly, a couple of files are copied over. A separate PR for SourceKittenFramework will be created to expose the necessary methods.