Cycle detection for pluginized component with non-core component (#216)
Because SourceKit does not provide implementation details of a property provided via `pluginExtension.foo`, we cannot implement an exact detection mechanism. The validator assumes if a property is required by the plugin extension, and the non-core component's dependnecy, yet provided by the core component, then this is a cycle. This is generally true.