Refactor file content filtering (#240)
* Refactor file content filtering

Modularize the file content filtering into separate filters. Each responsible for a specific type of DI unit. This allows for better composition among the filters.

The primary goal of this is to enable a separate pass of parsing component extensions, after all the components have been parsed out. Since the component extension filter can only be applied if we know the name of the components, a separate pass is needed.

Added the `ComponentExtensionFilter` and `ComponentExtensionsFilterTask`.

* Consolidate regex creation for class and protocol inheritances

* Consolidate keyword and regular expression based filter into base class