Significantly improve compile performance of generated code
The following commit makes the changes listed below to the generated
code to significantly improve compile speed of the generated code.

On an M1 Max 64GB MacBookPro, with Xcode 13.3 the following improvements
were measured:

 - x86_64 target: 350% faster compile time (~86s -> ~25s)
 - arm64 target: 1600% faster compile time (~360s -> 22s)

Clearly the arm64 target before these changes was a significant
regression from the x86_64 compile times. Thus, alternatively you could
think of this patch as:

 - Fixing an arm64 compile regression with Swift 5.6
 - Improving compile time of generated code by 350%

The changes made to the generated code include:

 1. Helper functions for `.parent` traversal chains
 2. Not using closures when registering factories, instead using named
    functions
 3. Removing class inheritance from the provider code. Instead, factory
    functions are written to perform the exact `.parent` traversal
    needed.
 4. De-duplicating all factory functions since .parent traversal is
    generic. Thus, any child nodes in the dependency tree that are the
    same level in subtree can use the same factory function.
 5. Reordering code so it is in dependency order.
 6. Adding a helper for registration calls, to remove lots of duplicate
    data structure traversal.
 7. Splitting registration into named functions no more than 50
    statements long. A single function with well over 10k registration
    calls was the primary cause of the arm64 compile time regression.

These changes not only reduced the compile time, but significantly
reduced the amount of code being compiled. In my tests that involved the
numbers below, the generated code went from around 18MB to 10MB.